+++ apps/api/src/health/health.module.ts
@@ -2,8 +2,10 @@ import { Module } from '@nestjs/common';
 
 import { HealthController } from './health.controller';
 import { HealthService } from './health.service';
+import { MailingModule } from '../shared/mailing.module';
 
 @Module({
+  imports: [MailingModule],
   controllers: [HealthController],
   providers: [HealthService],
 })